+++ src/main/java/mdlaf/animation/MaterialMouseHover.java
@@ -4,7 +4,8 @@ import java.awt.event.MouseListener;
 import java.awt.event.MouseMotionListener;
 
 /**
- * This is only an Tag Interface
+ * Tag interface, all material-ui-swing mouse hover should be implement this interface
+ *
  * @author https://github.com/vincenzopalazzo
  */
 public interface MaterialMouseHover extends MouseListener, MouseMotionListener { }