+++ src/main/java/mdlaf/animation/MaterialUIMovement.java
@@ -34,10 +34,20 @@ public class MaterialUIMovement {
 
 	private MaterialUIMovement () {}
 
+	/**
+	 * @deprecated This method will be deprecated inside the version 1.2 of the library.
+	 * you can {@see getMovement method}
+	 */
+	@Deprecated
 	public static void add (JComponent c, Color fadeTo, int steps, int interval) {
 		new MaterialUITimer (c, fadeTo, steps, interval);
 	}
 
+	/**
+	 * @deprecated This method will be deprecated inside the version 1.2 of the library.
+	 * you can {@see getMovement method}
+	 */
+	@Deprecated
 	public static void add (JComponent c, Color fadeTo) {
 		add (c, fadeTo, 5, 1000 / 60);
 	}
@@ -51,14 +61,17 @@ public class MaterialUIMovement {
 	}
 
 	/**
-	 * Thie method create a new effect mouse hover static, not create a wake effect
+	 * This method create a new effect mouse hover static, not create a wake effect
 	 * This method is util in all component for click, an example: The button for JSpinner, JCombobox, JScroolbar
 	 * param c is the component
 	 * param colorEffect is the color for effect, (For using color, look the MaterialColors class)
 	 * param colorOnClick is the color of the component on click on it
 	 * return a new MaterialUIStaticMovement this class implement a only MouseListner for moment
-	 * author https://github.com/vincenzopalazzo
+	 * @author https://github.com/vincenzopalazzo
+	 * @deprecated This method will be removed inside the version 1.2, because with the new MaterialButtonApi, you can create
+	 * a personal instance of the button, with an personal Listener, {@see MaterialButtonUI}
 	 */
+	@Deprecated
 	public static MaterialMouseHover getStaticMovement(JComponent c, Color colorEffect, Color colorOnClick){
 		if(c == null || colorEffect == null || colorOnClick == null){
 			throw new IllegalArgumentException("Che input arguments is/are null");