+++ src/test/java/integration/gui/mock/component/TableModelSecondPanel.java
@@ -72,5 +72,7 @@ public class TableModelSecondPanel extends AbstractTableModel {
             default: return null;
         }
     }
+
+
 }